utils: Add __init__.py to expose usable methods from utility class
Also, to indicate to Python to treat this as a directory containing utility packages which we might have in the future.
